chore:modified route
created separate files for each route
created separate files for each controller
[delivered #160850968]
