--- UI/debug.log
@@ -1,61 +1,0 @@
-[0904/121743.697:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0905/044557.193:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0905/052051.609:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0905/125603.648:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/100752.777:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/101752.779:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/101752.782:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/101810.398:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/101831.516:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/102810.407:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/102810.407:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/103310.413:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/103921.383:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/104810.422:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/110310.442:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/193617.924:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/141857.266:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/144124.263:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/144129.461:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/144138.897:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/151247.967:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/151454.516:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154106.684:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154128.595:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154206.379:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154209.379:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154211.939:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154230.653:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154244.317:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154353.079:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154404.096:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154408.467:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154416.363:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154418.014:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154419.564:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154421.030:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154427.160:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154429.496:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154436.984:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154439.317:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154733.138:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154735.643:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154738.288:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154740.023:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154745.053:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154747.139:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154749.710:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154752.338:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154802.890:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154804.419:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154805.699:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154807.219:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154808.698:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154809.935:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154811.193:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154812.352:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154820.685:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154826.435:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154839.131:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0906/154851.007:ERROR:crash_report_database_win.cc(428)] unexpected header
-[0907/032753.463:ERROR:crash_report_database_win.cc(428)] unexpected header
