+++ server/app.js
@@ -4,7 +4,6 @@ import myRoute from './routes/get';
 import route from './routes/create';
 import router from './routes/modify';
 import myRouter from './routes/delete';
-import rout from './routes/validat';
 const app = express();
 const port = process.env.PORT || 5000;
 
@@ -17,13 +16,14 @@ app.get('/',  (req, res) => {
     }) 
 })
 
-app.use('/api/v1', rout)
 app.use('/api/v1', myRoute)
 app.use('/api/v1', route)
 app.use('/api/v1', router)
 app.use('/api/v1', myRouter)
+
+ //added route to catch all errors
 app.use('*',  (req, res) => {
-    res.status(404).send({       //added route to catch all error
+    res.status(404).send({      
         status:"bad request"
     }) 
 })
