+++ server/routes/get.js
@@ -1,17 +1,11 @@
 import express from "express";
-import MyController from './../controllers/myControl';
+import MyController from './../controllers/getControl';
 const router = express.Router();
-import foodValidate from './../middleware/validate';
 
 router.route('/foods')
     .get(MyController.getAllFoods)
-//     .post(foodValidate.createOrderValidator,MyController.createNewOrder);
-
  router.route('/foods/:foodID')
   .get(MyController.getSpecificFoods)
-//     .put(foodValidate.modifyOrderValidator, MyController.modifyInput)
-//     .delete(MyController.deleteSpecificFoods)
-
